Gujcet Score changed from int to float Backend

+++ APTITUDE/APTITUDETEST.Core/Entities/Users/User.cs
@@ -37,7 +37,7 @@ namespace APTITUDETEST.Core.Entities.Users
         [ForeignKey("MasterTechnologies")]
         public int? TechnologyInterestedIn { get; set; }
         public int? ACPCMeritRank { get; set; }
-        public int? GUJCETScore { get; set; }
+        public float? GUJCETScore { get; set; }
         public float? JEEScore { get; set; }
         public int? Gender { get; set; }
         public int? PreferedLocation { get; set; }
